+++ crates/lib/src/errors/kind.rs
@@ -36,6 +36,8 @@ pub enum GDErrorKind {
     JsonParse,
     /// Couldn't parse a value.
     TypeParse,
+    /// Couldn't find the host specified.
+    HostLookup,
 }
 
 impl GDErrorKind {