+++ crates/lib/src/protocols/types.rs
@@ -19,6 +19,7 @@ pub enum ProprietaryProtocol {
     FFOW,
     JC2M,
     Savage2,
+    Eco,
     Mindustry,
 }
 
@@ -54,6 +55,8 @@ pub enum GenericResponse<'a> {
     JC2M(&'a crate::games::jc2m::Response),
     #[cfg(feature = "games")]
     Savage2(&'a crate::games::savage2::Response),
+    #[cfg(feature = "games")]
+    Eco(&'a crate::games::eco::Response),
 }
 
 /// All player types
@@ -71,6 +74,8 @@ pub enum GenericPlayer<'a> {
     TheShip(&'a crate::games::theship::TheShipPlayer),
     #[cfg(feature = "games")]
     JCMP2(&'a crate::games::jc2m::Player),
+    #[cfg(feature = "games")]
+    Eco(&'a crate::games::eco::Player),
 }
 
 pub trait CommonResponse {