+++ CHANGELOG.md
@@ -18,9 +18,14 @@ Games:
 - Added a valve protocol query example.
 - Made all of Just Cause 2: Multiplayer Response and Player fields public.
 - [Mindustry](https://mindustrygame.github.io/) support.
+- Eco support (by @CosminPerRam).
+
+Crate:
+- Changed the serde feature to only enable serde derive for some types: serde and serde_json is now a dependecy by default.
 
 Protocols:
 - Added the unreal2 protocol and its associated games: Darkest Hour, Devastation, Killing Floor, Red Orchestra, Unreal Tournament 2003, Unreal Tournament 2004 (by @Douile).
+- Added HTTPClient to allow use of HTTP(S) (and JSON) APIs (by @CosminPerRam & @Douile).
 
 Crate:
 - Added a `packet_capture` feature to capture the raw packets sent and received by the socket (by @Douile).